V1-LocalSolution/requirements.txt
mysql-connector-python==8.0.28
protobuf==3.19.3
python-decouple==3.5
